Eliminate file ledger_tokens.rs
