+++ aries_vcx/src/libindy/utils/mod.rs
@@ -11,7 +11,6 @@ pub mod signus;
 pub mod wallet;
 pub mod pool;
 pub mod crypto;
-pub mod ledger_tokens;
 pub mod cache;
 pub mod logger;
